Separate upgrading from LevelUpUI, use new upgrading classes
